+++ lib/nd_disposable.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/foundation.dart';
 
 abstract class NDDisposable {
+  @mustCallSuper
   void dispose();
 
   @visibleForTesting
-  bool get isDisposed;
+  bool get isDisposed => throw UnimplementedError(
+      'isDisposed is not implemented by ${runtimeType.toString()}');
 }