+++ lib/nd_disposable_wrapper.dart
@@ -1,15 +1,13 @@
+//
+// nd_disposable_wrapper.dart
+// nd_core_utils
+//
+// Created by Nguyen Duc Hiep on Tue May 03 2022.
+// Copyright © 2022 Nguyen Duc Hiep. All rights reserved.
+//
+
+import 'package:nd_core_utils/nd_auto_disposable.dart';
 import 'package:nd_core_utils/nd_disposable.dart';
 
-class NDDisposableWrapper<T extends NDDisposable> {
-  T? get value => _value;
-  void reset() => value = null;
-  set value(T? value) {
-    _value?.dispose();
-    _value = value;
-  }
-
-  NDDisposableWrapper(T? value) : _value = value;
-
-  // Privates
-  T? _value;
-}
+@Deprecated('Use NDAutoDisposable instead. Will be removed in 1.1')
+typedef NDDisposableWrapper<T extends NDDisposable> = NDAutoDisposable<T>;