+++ lib/nd_variable_wrapper.dart
@@ -5,6 +5,7 @@
 //  Created by Nguyen Duc Hiep on 01/12/2021.
 //
 
+@Deprecated('Idiot idea, will be removed in 1.1')
 class NDVariableWrapper<T> {
   T value;
   NDVariableWrapper(T _value) : value = _value;